Use bech32 for node addresses
