+++ radixdlt-core/radixdlt/src/integration/java/com/radixdlt/integration/mempool/MempoolRelayTest.java
@@ -21,6 +21,7 @@ import com.google.inject.multibindings.ProvidesIntoSet;
 import com.radixdlt.application.TokenUnitConversions;
 import com.radixdlt.chaos.mempoolfiller.MempoolFillerModule;
 import com.radixdlt.chaos.mempoolfiller.MempoolFillerUpdate;
+import com.radixdlt.client.Address;
 import com.radixdlt.ledger.LedgerAccumulator;
 import com.radixdlt.ledger.SimpleLedgerAccumulatorAndVerifier;
 import com.radixdlt.ledger.VerifiedTxnsAndProof;
@@ -68,7 +69,6 @@ import com.radixdlt.statecomputer.checkpoint.MockedGenesisModule;
 import com.radixdlt.store.DatabaseEnvironment;
 import com.radixdlt.store.DatabaseLocation;
 import com.radixdlt.store.berkeley.BerkeleyLedgerEntryStore;
-import com.radixdlt.utils.Base58;
 
 import java.util.Collection;
 import java.util.Comparator;
@@ -193,7 +193,7 @@ public class MempoolRelayTest {
 					bind(View.class).annotatedWith(EpochCeilingView.class).toInstance(View.of(100));
 					bind(MempoolConfig.class).toInstance(MempoolConfig.of(100L, 100L, 10L, 10L, 10));
 					bindConstant().annotatedWith(DatabaseLocation.class)
-						.to(folder.getRoot().getAbsolutePath() + "/" + Base58.toBase58(ecKeyPair.getPublicKey().getBytes()));
+						.to(folder.getRoot().getAbsolutePath() + "/" + Address.ofValidator(ecKeyPair.getPublicKey()));
 				}
 			}
 		);
