+++ radixdlt-core/radixdlt/src/integration/java/com/radixdlt/integration/recovery/OneNodeAlwaysAliveSafetyTest.java
@@ -17,6 +17,7 @@
 
 package com.radixdlt.integration.recovery;
 
+import com.radixdlt.client.Address;
 import com.radixdlt.ledger.LedgerAccumulator;
 import com.radixdlt.ledger.SimpleLedgerAccumulatorAndVerifier;
 import com.radixdlt.ledger.VerifiedTxnsAndProof;
@@ -80,7 +81,6 @@ import com.radixdlt.store.DatabaseEnvironment;
 import com.radixdlt.store.DatabaseLocation;
 import com.radixdlt.store.berkeley.BerkeleyLedgerEntryStore;
 import com.radixdlt.sync.messages.local.LocalSyncRequest;
-import com.radixdlt.utils.Base58;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -213,7 +213,7 @@ public class OneNodeAlwaysAliveSafetyTest {
 					bind(View.class).annotatedWith(EpochCeilingView.class).toInstance(View.of(88L));
 					bind(MempoolConfig.class).toInstance(MempoolConfig.of(10L, 10L));
 					bindConstant().annotatedWith(DatabaseLocation.class)
-						.to(folder.getRoot().getAbsolutePath() + "/" + Base58.toBase58(ecKeyPair.getPublicKey().getBytes()));
+						.to(folder.getRoot().getAbsolutePath() + "/" + Address.ofValidator(ecKeyPair.getPublicKey()));
 				}
 
 				@ProvidesIntoSet
